Uploading finished product
We’d only really uploaded some of the core files previously for some
reason, so I’m taking the opportunity to push everything that went into
the app (like we should have done in the first place). The files
index.html and those under /www/js, /www/templates, and /phpstuff
represent the majority of our work.